commenting and touching up #1

+++ Ad4m's Game/Assets/Scripts/PlayerAttributeController.cs
@@ -1,3 +1,5 @@
+//this is for all upgrades, attributes of the player. even upgrades will be kept here then will be used from the scripts using getcomponent.
+
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -7,5 +9,5 @@ public class PlayerAttributeController : MonoBehaviour
     public float playerBulletDamage = 20f;
     public float playerMeleeDamage = 40f;
     public float reloadSpeed = 1f;
-    public float shotAmount = 3f;
+    public int shotAmount = 3;
 }
